Put ticker class and pp_colors in hpp file in namespace and remove helper files
